+++ CHANGELOG.md
@@ -6,6 +6,9 @@ CHANGELOG
 - [sdk/python] Gracefully handle monitor shutdown in the python runtime without exiting the process. 
   [#6249](https://github.com/pulumi/pulumi/pull/6249)
 
+- [codegen/python] Generate accessors for provider resource properties.
+  [#6249](https://github.com/pulumi/pulumi/pull/6260)
+
 ## 2.20.0 (2021-02-03)
 
 - [sdk/python] Fix `Output.from_input` to unwrap nested output values in input types (args classes), which addresses